Correction from Nelson Beebe: argument to functions such as isspace() must be an int, not a char.  This could cause sign-extension problems.

+++ loadparm.c
@@ -479,11 +479,12 @@ static int strwicmp(char *psz1, char *psz2)
    /* sync the strings on first non-whitespace */
    while (1)
    {
-      while (isspace(*psz1))
+      while (isspace((int) *psz1))
          psz1++;
-      while (isspace(*psz2))
+      while (isspace((int) *psz2))
          psz2++;
-      if (toupper(*psz1) != toupper(*psz2) || *psz1 == '\0' || *psz2 == '\0')
+      if (toupper((int) *psz1) != toupper((int) *psz2)
+	  || *psz1 == '\0' || *psz2 == '\0')
          break;
       psz1++;
       psz2++;
